chore(wmg): move InfoPanel
