+++ frontend/src/views/WheresMyGeneV2/components/InfoPanel/components/ExpressedInCells/style.ts
