+++ frontend/src/views/WheresMyGeneV2/components/InfoPanel/components/Legend/index.tsx
