+++ frontend/src/views/WheresMyGeneV2/components/InfoPanel/components/Legend/style.ts
