+++ frontend/src/views/WheresMyGeneV2/components/Main/index.tsx
@@ -33,8 +33,8 @@ import CellInfoSideBar from "src/views/WheresMyGeneV2/components/CellInfoSideBar
 import Filters from "src/views/WheresMyGeneV2/components/Filters";
 import GeneInfoSideBar from "src/components/GeneInfoSideBar";
 
-import InfoPanel from "src/views/WheresMyGene/components/InfoPanel";
-import Legend from "src/views/WheresMyGene/components/InfoPanel/components/Legend";
+import InfoPanel from "src/views/WheresMyGeneV2/components/InfoPanel";
+import Legend from "src/views/WheresMyGeneV2/components/InfoPanel/components/Legend";
 import Loader from "src/views/WheresMyGene/components/Loader";
 import {
   StyledBannerContainer,
